requirements.txt
flask==3.0.3
flask-cors==4.0.0
pyodbc==5.1.0
sqlalchemy==2.0.30
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2
anthropic==0.26.0
python-dotenv==1.0.1
requests==2.31.0
flask-jwt-extended==4.6.0
bcrypt==4.1.2